refactor: 클래스간 구분을 위해 패키지 생성
