+++ src/main/java/lotto/LottoApplication.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.controller.LottoApplicationController;
+import lotto.domain.Repeater;
+
 public class LottoApplication {
 
 	public static void main(String[] args) {
