+++ src/main/java/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Lotto {
 	private final LottoNumbers lottoNumbers;
