+++ src/main/java/lotto/domain/Quantity.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Quantity {
 	private int totalLottoQuantity;
