+++ src/main/java/lotto/domain/Repeater.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Repeater {
 	private static boolean repeater;
