+++ src/main/java/lotto/domain/WinnerCount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class WinnerCount {
 	private int winnerCount;
