+++ src/main/java/lotto/domain/WinningAmount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class WinningAmount {
 	private final int winningAmount;
