+++ src/main/java/lotto/domain/WinningInformation.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class WinningInformation {
 	private final MatchedNumber matchedNumber;
